requirements.txt
tensorflow==2.17.1
fastapi
uvicorn
mlflow
pytest
pillow
numpy
httpx
python-multipart
prometheus-fastapi-instrumentator
